Safari-609.1.11.1/LayoutTests/fast/scrolling/unconstrained-scrollto-expected.txt

PASS scroller.scrollTop is -150
PASS successfullyParsed is true

TEST COMPLETE
